Refactor Vim configuration (2023)
This is a complete refactor of all of my Vim configuration to a simpler,
cleaner and more efficient Vim experience. This commit also improves the
microline plugin by introducing the use of `synIDtrans` to grab the
color attributes of the current color scheme to create a nicer looking
inverse colors effect for the sections and separators. Finally, I added
the gruvbox colorscheme for a more modern, nicer look & feel.