Replace kitty theme scripts with darkmode-listener
Thanks to bouk's utility, I can now respond to changes in the system's
appearance, so I renamed the ChangeHighlight function to
RefreshColorscheme and called it in a SigUSR1 autocmd, which essentially
replaces the manual kitty-* scripts. I also had to update microline to
support the SigUSR1 autocmd in addition to the Colorscheme autocmd.

https://github.com/bouk/dark-mode-notify